refactor: divvy up the db/batch impls into their own mods
Closes STOR-403
