+++ syncstorage-mysql/src/db/batch_impl.rs
@@ -1,6 +1,7 @@
 use base64::Engine;
 use std::collections::HashSet;
 
+use async_trait::async_trait;
 use diesel::{
     self,
     dsl::sql,
@@ -11,181 +12,182 @@ use diesel::{
     ExpressionMethods, OptionalExtension, QueryDsl,
 };
 use diesel_async::RunQueryDsl;
-use syncstorage_db_common::{params, results, UserIdentifier, BATCH_LIFETIME};
+use syncstorage_db_common::{params, results, BatchDb, Db, UserIdentifier, BATCH_LIFETIME};
 
 use super::{
     schema::{batch_upload_items, batch_uploads},
-    DbError, DbResult, MysqlDb,
+    MysqlDb,
 };
+use crate::{DbError, DbResult};
 
 const MAX_TTL: i32 = 2_100_000_000;
 
 const MAX_BATCH_CREATE_RETRY: u8 = 5;
 
-pub async fn create(
-    db: &mut MysqlDb,
-    params: params::CreateBatch,
-) -> DbResult<results::CreateBatch> {
-    let user_id = params.user_id.legacy_id as i64;
-    let collection_id = db.get_collection_id(&params.collection).await?;
-    // Careful, there's some weirdness here!
-    //
-    // Sync timestamps are in seconds and quantized to two decimal places, so
-    // when we convert one to a bigint in milliseconds, the final digit is
-    // always zero. But we want to use the lower digits of the batchid for
-    // sharding writes via (batchid % num_tables), and leaving it as zero would
-    // skew the sharding distribution.
-    //
-    // So we mix in the lowest digit of the uid to improve the distribution
-    // while still letting us treat these ids as millisecond timestamps.  It's
-    // yuck, but it works and it keeps the weirdness contained to this single
-    // line of code.
-    let mut batch_id = db.timestamp().as_i64() + (user_id % 10);
-    // Occasionally batch_ids clash (usually during unit testing), so also
-    // retry w/ increments
-    for i in 1..=MAX_BATCH_CREATE_RETRY {
-        let result = insert_into(batch_uploads::table)
-            .values((
-                batch_uploads::batch_id.eq(&batch_id),
-                batch_uploads::user_id.eq(&user_id),
-                batch_uploads::collection_id.eq(&collection_id),
-            ))
-            .execute(&mut db.conn)
-            .await;
-        match result {
-            Ok(_) => break,
-            Err(DieselError::DatabaseError(UniqueViolation, _)) => {
-                if i == MAX_BATCH_CREATE_RETRY {
-                    return Err(DbError::conflict());
+#[async_trait(?Send)]
+impl BatchDb for MysqlDb {
+    type Error = DbError;
+
+    async fn create_batch(
+        &mut self,
+        params: params::CreateBatch,
+    ) -> DbResult<results::CreateBatch> {
+        let user_id = params.user_id.legacy_id as i64;
+        let collection_id = self.get_collection_id(&params.collection).await?;
+        // Careful, there's some weirdness here!
+        //
+        // Sync timestamps are in seconds and quantized to two decimal places, so
+        // when we convert one to a bigint in milliseconds, the final digit is
+        // always zero. But we want to use the lower digits of the batchid for
+        // sharding writes via (batchid % num_tables), and leaving it as zero would
+        // skew the sharding distribution.
+        //
+        // So we mix in the lowest digit of the uid to improve the distribution
+        // while still letting us treat these ids as millisecond timestamps.  It's
+        // yuck, but it works and it keeps the weirdness contained to this single
+        // line of code.
+        let mut batch_id = self.timestamp().as_i64() + (user_id % 10);
+        // Occasionally batch_ids clash (usually during unit testing), so also
+        // retry w/ increments
+        for i in 1..=MAX_BATCH_CREATE_RETRY {
+            let result = insert_into(batch_uploads::table)
+                .values((
+                    batch_uploads::batch_id.eq(&batch_id),
+                    batch_uploads::user_id.eq(&user_id),
+                    batch_uploads::collection_id.eq(&collection_id),
+                ))
+                .execute(&mut self.conn)
+                .await;
+            match result {
+                Ok(_) => break,
+                Err(DieselError::DatabaseError(UniqueViolation, _)) => {
+                    if i == MAX_BATCH_CREATE_RETRY {
+                        return Err(DbError::conflict());
+                    }
+                    batch_id += 1;
                 }
-                batch_id += 1;
+                Err(e) => return Err(e.into()),
             }
-            Err(e) => return Err(e.into()),
         }
+
+        do_append(self, batch_id, params.user_id, collection_id, params.bsos).await?;
+        Ok(results::CreateBatch {
+            id: encode_id(batch_id),
+            size: None,
+        })
     }
 
-    do_append(db, batch_id, params.user_id, collection_id, params.bsos).await?;
-    Ok(results::CreateBatch {
-        id: encode_id(batch_id),
-        size: None,
-    })
-}
+    async fn validate_batch(&mut self, params: params::ValidateBatch) -> DbResult<bool> {
+        let batch_id = decode_id(&params.id)?;
+        // Avoid hitting the db for batches that are obviously too old.  Recall
+        // that the batchid is a millisecond timestamp.
+        if (batch_id + BATCH_LIFETIME) < self.timestamp().as_i64() {
+            return Ok(false);
+        }
 
-pub async fn validate(db: &mut MysqlDb, params: params::ValidateBatch) -> DbResult<bool> {
-    let batch_id = decode_id(&params.id)?;
-    // Avoid hitting the db for batches that are obviously too old.  Recall
-    // that the batchid is a millisecond timestamp.
-    if (batch_id + BATCH_LIFETIME) < db.timestamp().as_i64() {
-        return Ok(false);
+        let user_id = params.user_id.legacy_id as i64;
+        let collection_id = self.get_collection_id(&params.collection).await?;
+        let exists = batch_uploads::table
+            .select(sql::<Integer>("1"))
+            .filter(batch_uploads::batch_id.eq(&batch_id))
+            .filter(batch_uploads::user_id.eq(&user_id))
+            .filter(batch_uploads::collection_id.eq(&collection_id))
+            .get_result::<i32>(&mut self.conn)
+            .await
+            .optional()?;
+        Ok(exists.is_some())
     }
 
-    let user_id = params.user_id.legacy_id as i64;
-    let collection_id = db.get_collection_id(&params.collection).await?;
-    let exists = batch_uploads::table
-        .select(sql::<Integer>("1"))
-        .filter(batch_uploads::batch_id.eq(&batch_id))
-        .filter(batch_uploads::user_id.eq(&user_id))
-        .filter(batch_uploads::collection_id.eq(&collection_id))
-        .get_result::<i32>(&mut db.conn)
-        .await
-        .optional()?;
-    Ok(exists.is_some())
-}
+    async fn append_to_batch(&mut self, params: params::AppendToBatch) -> DbResult<()> {
+        let exists = self
+            .validate_batch(params::ValidateBatch {
+                user_id: params.user_id.clone(),
+                collection: params.collection.clone(),
+                id: params.batch.id.clone(),
+            })
+            .await?;
 
-pub async fn append(db: &mut MysqlDb, params: params::AppendToBatch) -> DbResult<()> {
-    let exists = validate(
-        db,
-        params::ValidateBatch {
+        if !exists {
+            return Err(DbError::batch_not_found());
+        }
+
+        let batch_id = decode_id(&params.batch.id)?;
+        let collection_id = self.get_collection_id(&params.collection).await?;
+        do_append(self, batch_id, params.user_id, collection_id, params.bsos).await?;
+        Ok(())
+    }
+
+    async fn get_batch(&mut self, params: params::GetBatch) -> DbResult<Option<results::GetBatch>> {
+        let is_valid = self
+            .validate_batch(params::ValidateBatch {
+                user_id: params.user_id,
+                collection: params.collection,
+                id: params.id.clone(),
+            })
+            .await?;
+        let batch = if is_valid {
+            Some(results::GetBatch { id: params.id })
+        } else {
+            None
+        };
+        Ok(batch)
+    }
+
+    async fn delete_batch(&mut self, params: params::DeleteBatch) -> DbResult<()> {
+        let batch_id = decode_id(&params.id)?;
+        let user_id = params.user_id.legacy_id as i64;
+        let collection_id = self.get_collection_id(&params.collection).await?;
+        diesel::delete(batch_uploads::table)
+            .filter(batch_uploads::batch_id.eq(&batch_id))
+            .filter(batch_uploads::user_id.eq(&user_id))
+            .filter(batch_uploads::collection_id.eq(&collection_id))
+            .execute(&mut self.conn)
+            .await?;
+        diesel::delete(batch_upload_items::table)
+            .filter(batch_upload_items::batch_id.eq(&batch_id))
+            .filter(batch_upload_items::user_id.eq(&user_id))
+            .execute(&mut self.conn)
+            .await?;
+        Ok(())
+    }
+
+    /// Commits a batch to the bsos table, deleting the batch when succesful
+    async fn commit_batch(
+        &mut self,
+        params: params::CommitBatch,
+    ) -> DbResult<results::CommitBatch> {
+        let batch_id = decode_id(&params.batch.id)?;
+        let user_id = params.user_id.legacy_id as i64;
+        let collection_id = self.get_collection_id(&params.collection).await?;
+        let timestamp = self.timestamp();
+        sql_query(include_str!("batch_commit.sql"))
+            .bind::<BigInt, _>(user_id)
+            .bind::<Integer, _>(&collection_id)
+            .bind::<BigInt, _>(&self.timestamp().as_i64())
+            .bind::<BigInt, _>(&self.timestamp().as_i64())
+            .bind::<BigInt, _>((MAX_TTL as i64) * 1000) // XXX:
+            .bind::<BigInt, _>(&batch_id)
+            .bind::<BigInt, _>(user_id)
+            .bind::<BigInt, _>(&self.timestamp().as_i64())
+            .bind::<BigInt, _>(&self.timestamp().as_i64())
+            .execute(&mut self.conn)
+            .await?;
+
+        self.update_collection(params::UpdateCollection {
             user_id: params.user_id.clone(),
+            collection_id,
             collection: params.collection.clone(),
-            id: params.batch.id.clone(),
-        },
-    )
-    .await?;
-
-    if !exists {
-        return Err(DbError::batch_not_found());
-    }
-
-    let batch_id = decode_id(&params.batch.id)?;
-    let collection_id = db.get_collection_id(&params.collection).await?;
-    do_append(db, batch_id, params.user_id, collection_id, params.bsos).await?;
-    Ok(())
-}
-
-pub async fn get(
-    db: &mut MysqlDb,
-    params: params::GetBatch,
-) -> DbResult<Option<results::GetBatch>> {
-    let is_valid = validate(
-        db,
-        params::ValidateBatch {
-            user_id: params.user_id,
-            collection: params.collection,
-            id: params.id.clone(),
-        },
-    )
-    .await?;
-    let batch = if is_valid {
-        Some(results::GetBatch { id: params.id })
-    } else {
-        None
-    };
-    Ok(batch)
-}
-
-pub async fn delete(db: &mut MysqlDb, params: params::DeleteBatch) -> DbResult<()> {
-    let batch_id = decode_id(&params.id)?;
-    let user_id = params.user_id.legacy_id as i64;
-    let collection_id = db.get_collection_id(&params.collection).await?;
-    diesel::delete(batch_uploads::table)
-        .filter(batch_uploads::batch_id.eq(&batch_id))
-        .filter(batch_uploads::user_id.eq(&user_id))
-        .filter(batch_uploads::collection_id.eq(&collection_id))
-        .execute(&mut db.conn)
-        .await?;
-    diesel::delete(batch_upload_items::table)
-        .filter(batch_upload_items::batch_id.eq(&batch_id))
-        .filter(batch_upload_items::user_id.eq(&user_id))
-        .execute(&mut db.conn)
-        .await?;
-    Ok(())
-}
-
-/// Commits a batch to the bsos table, deleting the batch when succesful
-pub async fn commit(
-    db: &mut MysqlDb,
-    params: params::CommitBatch,
-) -> DbResult<results::CommitBatch> {
-    let batch_id = decode_id(&params.batch.id)?;
-    let user_id = params.user_id.legacy_id as i64;
-    let collection_id = db.get_collection_id(&params.collection).await?;
-    let timestamp = db.timestamp();
-    sql_query(include_str!("batch_commit.sql"))
-        .bind::<BigInt, _>(user_id)
-        .bind::<Integer, _>(&collection_id)
-        .bind::<BigInt, _>(&db.timestamp().as_i64())
-        .bind::<BigInt, _>(&db.timestamp().as_i64())
-        .bind::<BigInt, _>((MAX_TTL as i64) * 1000) // XXX:
-        .bind::<BigInt, _>(&batch_id)
-        .bind::<BigInt, _>(user_id)
-        .bind::<BigInt, _>(&db.timestamp().as_i64())
-        .bind::<BigInt, _>(&db.timestamp().as_i64())
-        .execute(&mut db.conn)
+        })
         .await?;
 
-    db.update_collection(user_id as u32, collection_id).await?;
-
-    delete(
-        db,
-        params::DeleteBatch {
+        self.delete_batch(params::DeleteBatch {
             user_id: params.user_id,
             collection: params.collection,
             id: params.batch.id,
-        },
-    )
-    .await?;
-    Ok(timestamp)
+        })
+        .await?;
+        Ok(timestamp)
+    }
 }
 
 pub async fn do_append(
@@ -296,12 +298,4 @@ fn decode_id(id: &str) -> DbResult<i64> {
     decoded
         .parse::<i64>()
         .map_err(|e| DbError::internal(format!("Invalid batch_id: {}", e)))
-}
-
-macro_rules! batch_db_method {
-    ($name:ident, $batch_name:ident, $type:ident) => {
-        pub async fn $name(&mut self, params: params::$type) -> DbResult<results::$type> {
-            batch::$batch_name(self, params).await
-        }
-    };
 }
